Corrected some warnings Added SHA3-512 hash to external components Added _CRT_SECURE_NO_WARNINGS compiler definition to jpeg9d

+++ include/MOHPC/Utilities/ModelRenderer.h
@@ -70,7 +70,7 @@ namespace MOHPC
 		struct Pose
 		{
 			SafePtr<const SkeletonAnimation> animation;
-			int32_t frameNum;
+			uintptr_t frameNum;
 			float weight;
 
 			Pose()
@@ -111,7 +111,7 @@ namespace MOHPC
 		 * @param FrameNumber : The frame number from the animation
 		 * @param Weight : The animation weight
 		 */
-		MOHPC_EXPORTS void SetMovementPose(const SkeletonAnimation* Animation, uint32_t PoseIndex, int32_t FrameNumber = 0, float Weight = 1.f);
+		MOHPC_EXPORTS void SetMovementPose(const SkeletonAnimation* Animation, uint32_t PoseIndex, uintptr_t FrameNumber = 0, float Weight = 1.f);
 
 		/**
 		 * Sets the model's action pose to a corresponding animation.
@@ -120,7 +120,7 @@ namespace MOHPC
 		 * @param FrameNumber : The frame number from the animation
 		 * @param Weight : The animation weight
 		 */
-		MOHPC_EXPORTS void SetActionPose(const SkeletonAnimation*, uint32_t PoseIndex, int32_t FrameNumber = 0, float Weight = 1.f);
+		MOHPC_EXPORTS void SetActionPose(const SkeletonAnimation*, uint32_t PoseIndex, uintptr_t FrameNumber = 0, float Weight = 1.f);
 
 		/** Increments the frame number of all poses. */
 		MOHPC_EXPORTS void AdvancePosesFrame();
